resnet-inference/requirements.txt
Pillow==7.2.0
minio==6.0.0
